remote-agent/pyproject.toml
[project]
name = "remote-agent"
version = "0.1.0"
description = "LiveKit AI agent: sees through the car camera and drives via voice commands"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[openai]~=1.4",
    "livekit-plugins-silero>=1.0",
    "python-dotenv>=1.0",
    "car-protocol",
]

[tool.uv.sources]
car-protocol = { workspace = true }
